+++ Purchases/Classes/Public/RCPurchaserInfo.m
@@ -12,6 +12,7 @@
 
 @property (nonatomic) NSDictionary<NSString *, NSDate *> *expirationDates;
 @property (nonatomic) NSSet<NSString *> *nonConsumablePurchases;
+@property (nonatomic) NSString *originalApplicationVersion;
 
 @end
 
@@ -57,6 +58,9 @@ static dispatch_once_t onceToken;
 
         NSDictionary<NSString *, id> *otherPurchases = subscriberData[@"other_purchases"];
         self.nonConsumablePurchases = [NSSet setWithArray:[otherPurchases allKeys]];
+
+        NSString *originalApplicationVersion = subscriberData[@"original_application_version"];
+        self.originalApplicationVersion = [originalApplicationVersion isKindOfClass:[NSNull class]] ? nil : originalApplicationVersion;
 
     }
     return self;