+++ Purchases/Classes/Public/RCPurchases.h
@@ -110,6 +110,10 @@ typedef void (^RCReceiveIntroEligibilityBlock)(NSDictionary<NSString *, RCIntroE
 - (void)checkTrialOrIntroductoryPriceEligibility:(NSArray<NSString *> *)productIdentifiers
                                       completion:(RCReceiveIntroEligibilityBlock)receiveEligibility;
 
+/**
+ Reads the App Store receipt and reads the original application version. Use this if RCPurchaserInfo.originalApplicationVersion is nil.
+ */
+- (void)updateOriginalApplicationVersion:(RCReceivePurchaserInfoBlock)receivePurchaserInfo;
 
 /**
  This version of the Purchases framework